backend/requirements.txt
flask>=2.3,<3.1
flask-cors>=4.0,<5.0
python-dotenv
huggingface_hub==0.23.0